+++ api/src/endware/index.ts
@@ -1,9 +1,7 @@
 import login from './login'
 import * as Report from './Report'
-import * as Archive from './Archive'
 
 export {
   login,
   Report,
-  Archive
 }