+++ api/src/middleware/index.ts
@@ -3,7 +3,6 @@ import verifyToken from "./verifyToken";
 import verifyVersion from "./verifyVersion";
 import verifyReportFileID from "./verifyReportFileID";
 import verifyReportFileID_min from "./verifyReportFileID_min";
-import verifyArchiveID from "./verifyArchiveID";
 
 export default { 
   verifyToken,
@@ -11,5 +10,4 @@ export default {
   verifyReportID,
   verifyReportFileID,
   verifyReportFileID_min,
-  verifyArchiveID
 };