+++ api/src/types/restrictions.ts
@@ -1,13 +1,15 @@
 export type restrictions = {
     private: boolean;
     whitelist: number[];
+    archive: boolean;
 }
 
 export const defaultRestrictions: restrictions = {
     private: true,
-    whitelist: []
+    whitelist: [],
+    archive: false
 }
 
-export const toRestrictions = (restrictions: string): restrictions => {return {...defaultRestrictions, ...JSON.parse(restrictions)}}
+export const toRestrictions = (restrictions: string | undefined): restrictions => {return restrictions != undefined ? {...defaultRestrictions, ...JSON.parse(restrictions)} : defaultRestrictions;}
 
 export default restrictions;